tg_raw_data_parser/requirements.txt
telethon
python-dotenv
loguru
uvicorn
fastapi